# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News aggregator API documentation",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/articles": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "articles resource",
                "operationId": "daffc9d151a0c7c956d1157d38e78859",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Global search keyword",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "published_at",
                        "in": "query",
                        "description": "The publish date of the article",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "The category of the article",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "The source of the article",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "number of articles per page",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "data": {
                                                "slug": "article1",
                                                "title": "article1",
                                                "description": "Laborum nesciunt ..",
                                                "content": "etur minus architect ..",
                                                "author": "r King",
                                                "category": "science",
                                                "source": "Newscred",
                                                "published_at": "2020-01-01"
                                            },
                                            "links": {
                                                "first": "articles?page=1",
                                                "next": "articles?page=2"
                                            },
                                            "meta": {
                                                "current_page": 1,
                                                "from": 1,
                                                "path": "/articles",
                                                "per_page": 10,
                                                "to": 10
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/feed": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "show feed",
                "operationId": "5f7ebc67d32a8e52b2e54ccbf3bf0fb1",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "number of articles per page",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "data": {
                                                "slug": "article1",
                                                "title": "article1",
                                                "description": "Laborum nesciunt ..",
                                                "content": "etur minus architect ..",
                                                "author": "r King",
                                                "category": "science",
                                                "source": "Newscred",
                                                "published_at": "2020-01-01"
                                            },
                                            "links": {
                                                "first": "articles?page=1",
                                                "next": "articles?page=2"
                                            },
                                            "meta": {
                                                "current_page": 1,
                                                "from": 1,
                                                "path": "/articles",
                                                "per_page": 10,
                                                "to": 10
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "user doesn't have preferences.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "user doesn't have preferences."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}